osd/scrub: extract ReplicaReservations into separate files
As a preliminary step before ReplicaReservations ownership
is moved to the scrubber's FSM.

No code changes in this commit (apart from required 'include's).
